requirements.txt
fastapi
uvicorn
jinja2
python-multipart
openai
pydantic
reportlab
python-docx